Move geo serializers
